feat: integrate allergens display in ItemDetail and update styles
