+++ package.json
@@ -57,7 +57,7 @@
     "@radix-ui/react-toggle": "^1.1.0",
     "@radix-ui/react-toggle-group": "^1.1.0",
     "@radix-ui/react-tooltip": "^1.1.4",
-    "@react-email/components": "^0.0.29",
+    "@react-email/components": "^0.0.30",
     "@sentry/nextjs": "^8",
     "@sindresorhus/slugify": "^2.2.1",
     "@stripe/stripe-js": "^5.2.0",
@@ -65,23 +65,23 @@
     "@t3-oss/env-nextjs": "^0.11.1",
     "@tailwindcss/container-queries": "^0.1.1",
     "@tanstack/match-sorter-utils": "^8.19.4",
-    "@tanstack/react-query": "^5.62.2",
-    "@tanstack/react-query-next-experimental": "^5.62.2",
+    "@tanstack/react-query": "^5.62.3",
+    "@tanstack/react-query-next-experimental": "^5.62.3",
     "@tanstack/react-table": "^8.20.5",
     "@types/canvas-confetti": "^1.6.4",
     "@types/papaparse": "^5.3.15",
     "@types/random-gradient": "^0.0.2",
     "@uiw/react-color": "^2.3.3",
-    "@uppy/aws-s3": "^4.1.2",
-    "@uppy/compressor": "^2.1.0",
-    "@uppy/core": "^4.2.3",
-    "@uppy/dashboard": "^4.1.2",
-    "@uppy/drag-drop": "^4.0.4",
-    "@uppy/file-input": "^4.0.3",
-    "@uppy/image-editor": "^3.2.0",
-    "@uppy/locales": "^4.3.0",
-    "@uppy/progress-bar": "^4.0.1",
-    "@uppy/react": "^4.0.3",
+    "@uppy/aws-s3": "^4.1.3",
+    "@uppy/compressor": "^2.1.1",
+    "@uppy/core": "^4.3.0",
+    "@uppy/dashboard": "^4.1.3",
+    "@uppy/drag-drop": "^4.0.5",
+    "@uppy/file-input": "^4.0.4",
+    "@uppy/image-editor": "^3.2.1",
+    "@uppy/locales": "^4.3.1",
+    "@uppy/progress-bar": "^4.0.2",
+    "@uppy/react": "^4.0.4",
     "@vercel/flags": "^2.6.3",
     "@vercel/toolbar": "^0.1.27",
     "canvas-confetti": "^1.9.3",
@@ -90,7 +90,7 @@
     "cmdk": "^1.0.4",
     "date-fns": "^4.1.0",
     "embla-carousel-react": "^8.5.1",
-    "extract-colors": "^4.1.0",
+    "extract-colors": "^4.1.1",
     "framer-motion": "12.0.0-alpha.1",
     "html2canvas": "^1.4.1",
     "input-otp": "^1.4.1",
@@ -99,24 +99,24 @@
     "lodash.isequal": "^4.5.0",
     "lodash.isobject": "^3.0.2",
     "lodash.transform": "^4.6.0",
-    "lucide-react": "^0.464.0",
+    "lucide-react": "^0.468.0",
     "lzutf8": "^0.6.3",
     "nanoid": "^5.0.9",
-    "next": "15.0.3",
+    "next": "15.0.4",
     "next-auth": "5.0.0-beta.24",
-    "next-axiom": "^1.6.0",
+    "next-axiom": "^1.7.0",
     "next-safe-action": "^7.9.9",
-    "next-themes": "^0.4.3",
+    "next-themes": "^0.4.4",
     "nuqs": "^2.2.3",
     "papaparse": "^5.4.1",
-    "posthog-js": "^1.194.3",
+    "posthog-js": "^1.194.5",
     "random-gradient": "^0.0.2",
-    "react": "19.0.0-rc-66855b96-20241106",
+    "react": "19.0.0",
     "react-aria": "^3.36.0",
     "react-contenteditable": "^3.3.7",
     "react-day-picker": "^9.4.1",
-    "react-dom": "19.0.0-rc-66855b96-20241106",
-    "react-hook-form": "^7.53.2",
+    "react-dom": "19.0.0",
+    "react-hook-form": "^7.54.0",
     "react-hot-toast": "^2.4.1",
     "react-joyride": "^3.0.0-6",
     "react-photo-view": "^1.2.6",
@@ -161,12 +161,12 @@
     "autoprefixer": "^10.4.20",
     "commitlint": "19.6.0",
     "eslint": "9.16.0",
-    "eslint-config-next": "15.0.3",
+    "eslint-config-next": "15.0.4",
     "eslint-plugin-react": "^7.37.2",
     "husky": "^9.1.7",
     "lint-staged": "^15.2.10",
     "postcss": "^8.4.49",
-    "prettier": "^3.4.1",
+    "prettier": "^3.4.2",
     "prettier-plugin-tailwindcss": "^0.6.9",
     "prisma": "^6.0.1",
     "react-email": "^3.0.3",
